+++ src/game/tests.rs
@@ -373,6 +373,7 @@ mod activation_timing;
 mod additional_costs;
 mod alternative_costs;
 mod attack_restrictions;
+mod aura_death;
 mod aura_upkeep;
 mod banding;
 mod blocking_relationship;